pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "whispr"
dynamic = ["version"]
description = 'A CLI tool to whisper your secrets between secure vault and your local environment'
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
keywords = [
    "cybersecurity",
    "mitre-attack-framework",
    "code-security",
    "no-plain-secrets",
    "whispr",
    "devsecops",
]
authors = [{ name = "Naren Yellavula", email = "naren.yellavula@gmail.com" }]

classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]

dependencies = [
    "click==8.1.7",
    "python-dotenv==1.0.1",
    "boto3==1.35.42",
    "PyYAML==6.0.2",
    "google-cloud-secret-manager==2.20.2",
    "structlog==24.4.0",
    "azure-keyvault==4.2.0",
    "azure-identity==1.19.0"
]

[project.urls]
Documentation = "https://github.com/narenaryan/whispr/blob/main/README.md"
Issues = "https://github.com/narenaryan/whispr/issues"
Source = "https://github.com/narenaryan/whispr"

[tool.hatch.version]
path = "src/whispr/__about__.py"

[tool.hatch.envs.default]
dependencies = ["coverage[toml]>=6.5", "pytest"]

[[tool.hatch.envs.all.matrix]]
python = ["3.8", "3.9", "3.10", "3.11", "3.12"]

[tool.hatch.envs.types]
dependencies = ["mypy>=1.0.0"]

[tool.coverage.run]
source_pkgs = ["whispr", "tests"]
branch = true
parallel = true
omit = ["src/whispr/__about__.py"]

[tool.coverage.paths]
whispr = ["src/whispr"]
tests = ["tests", "*/whispr/tests"]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]

[project.scripts]
whispr = "whispr.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/whispr"]

[tool.pyright]
venvPath = "/opt/homebrew/Caskroom/miniconda"
venv = "whispr"
